Move cxxtests to the test/ folder and port some of the old tests to cxxtests.
