+++ test/zcm/wscript
@@ -44,30 +44,6 @@ def build(ctx):
                 rpath = ctx.env.RPATH_zcm,
                 install_path = None)
 
-    ctx.program(target = 'forking',
-                use = 'default zcm',
-                source = 'forking.cpp',
-                rpath = ctx.env.RPATH_zcm,
-                install_path = None)
-
-    ctx.program(target = 'forking2',
-                use = 'default zcm',
-                source = 'forking2.cpp',
-                rpath = ctx.env.RPATH_zcm,
-                install_path = None)
-
-    ctx.program(target = 'flushing',
-                use = 'default zcm',
-                source = 'flushing.cpp',
-                rpath = ctx.env.RPATH_zcm,
-                install_path = None)
-
-    ctx.program(target = 'logtest',
-                use = 'default zcm',
-                source = 'logtest.cpp',
-                rpath = ctx.env.RPATH_zcm,
-                install_path = None)
-
     ctx.program(target = 'trackers',
                 use = 'default zcm testzcmtypes_cpp',
                 source = 'tracker_test.cpp',
