+++ zcm/wscript
@@ -92,8 +92,7 @@ def build(ctx):
     if ctx.env.USING_JULIA:
         ctx.recurse('julia')
 
-    if ctx.env.USING_CXXTEST:
-        ctx.cxxtest(use = ['zcm', 'testzcmtypes'])
+
 
 def package_embedded(tsk):
     gen = tsk.generator
